Menambahkan dan Mencoba perintah untuk proses PUT

+++ src/main/java/ws/a/praktikum3/ProductServiceController.java
@@ -5,6 +5,7 @@
  */
 package ws.a.praktikum3;
 
+import exception.ProductNotfoundException;
 import java.util.HashMap;
 import java.util.Map;
 import model.Product;
@@ -44,10 +45,13 @@ public class ProductServiceController {
     
     @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
     public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product){
-        productRepo.remove(id);
-        product.setId(id);
-        productRepo.put(id, product);
-        return new ResponseEntity<>("Product is updated successfully", HttpStatus.OK);
+        
+        if (!productRepo.containsKey(id)) throw new ProductNotfoundException();
+        
+            productRepo.remove(id);
+            product.setId(id);
+            productRepo.put(id, product);
+            return new ResponseEntity<>("Product is updated successfully", HttpStatus.OK);
     }
     
     @RequestMapping(value = "/products", method = RequestMethod.POST)
